Move start function to the start code file
